config.py
"""
Fichier listant toutes les dépendances du projet
"""
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv('SECRET_KEY') or 'a7d2ez@z/9dpi7hjde2'
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL') or 'sqlite:///../datas/recom-films.db'
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    JWT_SECRET_KEY = os.getenv('JWT_SECRET') or 'a7d2ez@z/9dpi7hjde2'