requirements.txt
numpy
pandas
scikit-learn
faiss-cpu
seaborn
matplotlib
requests
beautifulsoup4
lxml
Flask==2.0.1
flask-cors
Werkzeug==2.0.1
SQLAlchemy==1.4.46
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
django==4.2.16